refactor: move logic into separate file
